Regroup file into src and extract nyrc config :file_cabinet:
