+++ .nycrc
@@ -1,0 +1,10 @@
+{
+  "all": true,
+  "cache": true,
+  "exclude": ["scripts"],
+  "check-coverage": true,
+  "lines": 100,
+  "statements": 100,
+  "functions": 100,
+  "branches": 100
+}
