+++ src/config/core.js
