+++ src/config/css-modules.js
