+++ src/config/react.js
