+++ test/no-async-callback.js
@@ -1,6 +1,6 @@
 import test from 'ava';
 import avaRuleTester from 'eslint-ava-rule-tester';
-import rule from '../rules/no-async-callback';
+import rule from '../src/rules/no-async-callback';
 
 const ruleTester = avaRuleTester(test, {
   parserOptions: {
